fix 32bit format string warnings
git-svn-id: http://subversion.ffado.org/ffado/trunk@1764 2be59082-3212-0410-8809-b0798e1608f0

+++ libffado/src/libstreaming/amdtp-oxford/AmdtpOxfordReceiveStreamProcessor.cpp
@@ -189,9 +189,12 @@ AmdtpOxfordReceiveStreamProcessor::processPacketHeader(unsigned char *data, unsi
             }
             m_dll_e2 += m_dll_c * err;
 
-            debugOutput(DEBUG_LEVEL_VERY_VERBOSE, "Generated TSP: %16"PRIu64" %"PRId64" %"PRId32" %"PRId64"\n", 
-                                             m_next_packet_timestamp, m_next_packet_timestamp-m_last_timestamp,
-                                             frames_in_tempbuffer,  ((length / sizeof (quadlet_t)) - 2) / m_dimension);
+            debugOutput(DEBUG_LEVEL_VERY_VERBOSE,
+                        "Generated TSP: %16"PRIu64" %"PRId64" %d %d\n", 
+                        m_next_packet_timestamp,
+                        m_next_packet_timestamp-m_last_timestamp,
+                        (int)frames_in_tempbuffer,
+                        (int)(((length / sizeof (quadlet_t)) - 2) / m_dimension));
         }
 
         // add the payload to the temporary ringbuffer
